Reorganized and added new Split Filters, and fixed WWW isssues
