+++ src_2/lib/urls/register.ts
@@ -7,6 +7,9 @@ import * as cheerio from 'cheerio';
 
 import {RestURL, safeJSON} from '@otter-co/ottlib';
 
+import * as ksData from '../scrape_profiles/kickstarter';
+import * as igData from '../scrape_profiles/indiegogo';
+
 export class ContactUsURL extends RestURL implements RestURL
 {
     public static url = "/v1/register/contact_us";
@@ -184,14 +187,14 @@ export class RegisterProjectURL extends RestURL implements RestURL
 
             case 'https://www.indiegogo.com/':
                 url = dUrl + url;
-                scrapeProfile = this.igPageIDs;
-                scrapeMetaFunc = (wd)=>({'nope':true});
+                scrapeProfile = igData.pageIDs;
+                scrapeMetaFunc = igData.metaDataFunc;
                 break;
 
             case 'https://www.kickstarter.com/':
                 url = dUrl + url;
-                scrapeProfile = this.ksPageIDs;
-                scrapeMetaFunc = this.ksMetaData;
+                scrapeProfile = ksData.pageIDs;
+                scrapeMetaFunc = ksData.metaDataFunc;
             break;
 
             default:
@@ -204,7 +207,7 @@ export class RegisterProjectURL extends RestURL implements RestURL
         console.log(rawWData);
 
         if(!webData.title.content)
-            return this.end(rest, {success: false, data: {web_data_error: true, rawWData}});
+            return this.end(rest, {success: false, data: {web_data_error: true}});
             
         let coupon_code = crypto.randomBytes(6).toString('base64');
         let unique_id = crypto.randomBytes(10).toString('base64').split('').slice(0,6).join('');
@@ -213,7 +216,7 @@ export class RegisterProjectURL extends RestURL implements RestURL
             unique_id,
             name: webData.title.content,
             owner: sessR.data[0].username,
-            platform: (dUrl === 'https://kickstarter.com/') ? 'kickstarter' : 'indiegogo',
+            platform: (dUrl === 'https://www.kickstarter.com/') ? 'kickstarter' : 'indiegogo',
             coupon_code,
             project_data: {
                 web_data: webData,
@@ -238,121 +241,6 @@ export class RegisterProjectURL extends RestURL implements RestURL
             let usrU = await userG.set({username: sessR.data[0].username}, {'$set':{level:'PO'}}).catch(err=>err);
             return this.end(rest, {success: true});
         }
-    }
-
-    public ksPageIDs = {
-        title: [
-            'meta[property="og:title"]',
-
-            'content'
-        ],
-        description: [
-            'meta[property="og:description"]',
-
-            'content'
-        ],
-        content: [
-            "div.full-description",
-
-            'text',
-            'html'
-        ],
-        stats: [
-            "#pledged",
-
-            'data-goal',
-            'data-percent-raised',
-            'data-pledged'
-        ],
-        mainImg: [
-            'meta[property="og:image"]',
-
-            'content'
-        ],
-        hours: [
-            'span[data-hours-remaining]',
-
-            'data-duration',
-            'data-end_time',
-            'text'
-        ],
-        funding: [
-            'span.money',
-
-            'text'
-        ]
-    };
-
-    public igPageIDs = {
-        title: [
-            'meta[property="og:title"]',
-
-            'content'
-        ],
-        description: [
-            'meta[name="description"]',
-
-            'content'
-        ],
-        content: [
-            'div[gogo-test="description"]',
-
-            'text',
-            'html'
-        ],
-        stats: [
-            'span[gogo-test="raised"]',
-
-            'text',
-        ],
-        percentRaised: [
-            'meta[name="sailthru.pct_funded"]',
-
-            'content'
-        ],
-        mainImg: [
-            'meta[property="og:image"]',
-
-            'content'
-        ],
-        hours: [
-            'meta[name="sailthru.displayed_days_left"]',
- 
-            'content'
-        ],
-        funding: [
-            'div[gogo-test="percent_funded"]',
-
-            'text',
-            'html'
-        ]
-    };
-
-    //Regardless of this evidence, I do not hate what I do. I'm just lazy today.
-    protected ksMetaData(wd): any
-    {
-        let retO: any = {};
-        try
-        {
-            let fund = wd.funding.text;
-
-            retO.funding = wd.funding.text.split( /(\$|\€|MX\$)/g )[2];
-            //retO.fundingTest = fund.split( /(\$|\€|MX\$)/g ).filter( el => !(el.contains('MX$') || el.contains('$') || el.contains('€')) );
-    
-            retO.raisedPercent = +wd.stats['data-percent-raised'];
-            retO.raised = +wd.stats['data-percent-raised'] * retO.funding;
-
-            retO.featured = false;
-            retO.explore = false;
-            retO.landing = false;
-            retO.social = false;
-        }
-        catch(e)
-        {
-            console.log(e);
-        }
-
-        return retO;
     }
 
     protected getURLData(data, dataT): any
